sibes_select allows convenient #filtering of a dataset by any number of logical filters. For example, this function can be used to easily filter timestamps in a range, as well as combine simple spatial and temporal filters. It accepts a character vector of R expressions that each return a logical vector (i.e., TRUE or FALSE). Each filtering condition is interpreted in the context of the dataset supplied, and used to filter for rows that satisfy each of the filter conditions. Users must make sure that the filtering variables exist in their dataset in order to avoid errors.

sibes_select allows convenient #filtering of a dataset by any number of logical filters. For example, this function can be used to easily filter timestamps in a range, as well as combine simple spatial and temporal filters. It accepts a character vector of R expressions that each return a logical vector (i.e., TRUE or FALSE). Each filtering condition is interpreted in the context of the dataset supplied, and used to filter for rows that satisfy each of the filter conditions. Users must make sure that the filtering variables exist in their dataset in order to avoid errors.

sibes_select(data, filters = c())

Arguments

data

A tibble, dataframe or similar containing the variables to be filtered.

filters

A character vector of filter expressions. An example might be "sampling_station_id < 1000". The filtering variables must be in the dataframe. The function will not explicitly check whether the filtering variables are present; This makes it flexible, allowing expressions such as "between(sampling_station_id, 2, 2000)", but also something to use at your own risk. A missing filter variables will result in an empty data frame.

Value

A dataframe filtered using the filters specified.

Examples

data_list <- get('SIBES_dataset')

data_m <- sibes_merge(Inputdata = data_list)
#> Joining with `by = join_by(sample_id)`

data_selected_3 <- sibes_select(data = data_m, filters =c(
"between(sampling_station_id, 320,371)", #Location
"between(date,as.Date(\"2010-01-01\",format=\"%Y-%m-%d\"),
 as.Date(\"2020-08-11\",format=\"%Y-%m-%d\"))"))#Date