sibes_species_search This function use keywords to search the species tibble. This is convenient if only parts of a name are known, or for example, when multiple species need to be selected with a shared keyword

sibes_species_search This function use keywords to search the species tibble. This is convenient if only parts of a name are known, or for example, when multiple species need to be selected with a shared keyword

sibes_species_search(Inputdata = NULL, Key = NULL)

Arguments

Inputdata

Input data for the function is the output format of sibes_load which is data list of tibbels including "species". If these names are not present, the function returns an error

Key

The keyword to search for as a string. Numeric values also entered as character

Value

Function returns the found search.

Examples

data_list <- get('SIBES_dataset')

sibes_species_search(data_list[["species"]], Key='57|34|67')
#>    sibes_id                   name short_name aphia_id taxonomic_group
#> 1        34  Lepidonotus squamatus     Lepsqu   130801      polychaete
#> 2        57              Abra alba     Abralb   141433         bivalve
#> 3        67        Magallana gigas     Maggig   836033         bivalve
#> 4       134     Diogenes pugilator     Diopug   107199       crustacea
#> 5       234    Hypereteone foliosa     Hypfol   152250      polychaete
#> 6       267             Paguroidea     Paguro   106687       crustacea
#> 7       341          Proceraea sp.     Prospe   129671      polychaete
#> 8       342          Phyllodocidae     Phyllo      931      polychaete
#> 9       343  Boccardia proboscidea     Bocpro   327249      polychaete
#> 10      344           Necora puber     Necpub   107398       crustacea
#> 11      346      Palaemon serratus    Palaser   107616       crustacea
#> 12      347        Caprella mutica     Capmut   146768       crustacea
#> 13      348 Palaemon macrodactylus     Palmac   181372       crustacea
#> 14      367     Hippolyte varians      Hopvar   107518       crustacea
#>    year_added weight_is_measured length_is_measured length_measuring_method_id
#> 1        1999               True              False                         NA
#> 2        1999               True               True                          6
#> 3        1999               True               True                          6
#> 4        1999               True              False                         NA
#> 5        2015               True              False                         NA
#> 6        2018               True               True                         NA
#> 7        2020               True              False                         NA
#> 8        2021               True              False                         NA
#> 9        2021               True              False                         NA
#> 10       2021               True               True                          3
#> 11       2021               True               True                          2
#> 12       2021               True              False                         NA
#> 13       2021               True               True                          2
#> 14       2022               True              False                          2
#>    min_shell_length_to_separate_flesh missing_afdm_method_id remarks
#> 1                                  NA                     NA        
#> 2                                   8                     12        
#> 3                                   8                     12        
#> 4                                  NA                     NA        
#> 5                                  NA                     12        
#> 6                                  NA                     NA        
#> 7                                  NA                     NA        
#> 8                                  NA                     NA        
#> 9                                  NA                     NA        
#> 10                                 NA                     NA        
#> 11                                 NA                     NA        
#> 12                                 NA                     NA        
#> 13                                 NA                     NA        
#> 14                                 NA                     NA