sibes_species_search This function use keywords to search the species tibble. This is convenient if only parts of a name are known, or for example, when multiple species need to be selected with a shared keyword

sibes_species_search This function use keywords to search the species tibble. This is convenient if only parts of a name are known, or for example, when multiple species need to be selected with a shared keyword

sibes_species_search(Inputdata = NULL, Key = NULL)

Arguments

Inputdata

Input data for the function is the output format of sibes_load which is data list of tibbels including "species". If these names are not present, the function returns an error

Key

The keyword to search for as a string. Numeric values also entered as character

Value

Function returns the found search.

Examples

data_list <- get('SIBES_dataset')

sibes_species_search(data_list[["species"]], Key='57|34|67')
#>    sibes_id                   name short_name aphia_id taxonomic_group
#>       <int>                 <char>     <char>    <int>          <char>
#> 1:       57              Abra alba     Abralb   141433         bivalve
#> 2:       67        Magallana gigas     Maggig   836033         bivalve
#> 3:      267             Paguroidea     Paguro   106687       crustacea
#> 4:      348 Palaemon macrodactylus     Palmac   181372       crustacea
#> 5:      343  Boccardia proboscidea     Bocpro   327249      polychaete
#> 6:      234    Hypereteone foliosa     Hypfol   152250      polychaete
#> 7:       34  Lepidonotus squamatus     Lepsqu   130801      polychaete
#> 8:      342          Phyllodocidae     Phyllo      931      polychaete
#> 9:      341          Proceraea sp.     Prospe   129671      polychaete
#>    taxonomic_indentification_level year_added occurance occupancy
#>                             <char>      <int>     <int>     <num>
#> 1:                         species       2000       397       0.8
#> 2:                         species       2000       174       0.3
#> 3:                     superfamily       2018         1       0.0
#> 4:                         species       2021         1       0.0
#> 5:                         species       2021         1       0.0
#> 6:                         species       2015       205       0.4
#> 7:                         species       2000         4       0.0
#> 8:                          family       2021        17       0.0
#> 9:                           genus       2020         8       0.0
#>    weight_is_measured length_measuring_method_id.
#>                <lgcl>                       <int>
#> 1:               TRUE                           6
#> 2:               TRUE                           6
#> 3:               TRUE                          NA
#> 4:               TRUE                           2
#> 5:               TRUE                          NA
#> 6:               TRUE                          NA
#> 7:               TRUE                          NA
#> 8:               TRUE                          NA
#> 9:               TRUE                          NA
#>    min_shell_length_to_separate_flesh missing_afdm_method_id..
#>                                 <int>                    <int>
#> 1:                                  8                       12
#> 2:                                  8                       12
#> 3:                                 NA                       NA
#> 4:                                 NA                       NA
#> 5:                                 NA                       NA
#> 6:                                 NA                       12
#> 7:                                 NA                       NA
#> 8:                                 NA                       NA
#> 9:                                 NA                       NA
#>                                                                       remarks
#>                                                                        <char>
#> 1:                                                                           
#> 2:                                                                           
#> 3: Juvenile or broken individuals. So far, only Pagurus bernardus was found. 
#> 4:                                                                           
#> 5:                                                                           
#> 6:                                                                           
#> 7:                                                                           
#> 8:                  Phyllodocidae below 10 mm are classified on family level.
#> 9:                                           Juvenile or damaged individuals.